[refact] Move some models into 'models' module
