+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2025-05-08T08:31:42Z",
+  "generated_at": "2025-07-07T09:49:01Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -81,7 +81,7 @@
       {
         "hashed_secret": "fb34629c9af1ed4045b5d6f287426276b2be3a1e",
         "is_verified": false,
-        "line_number": 44,
+        "line_number": 43,
         "type": "Secret Keyword",
         "verified_result": null
       }
