+++ src/models/feature_snapshot.rs
@@ -12,14 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use murmur3::murmur3_32;
+
 use crate::entity::Entity;
 use crate::value::Value;
 use crate::Feature;
 
-use super::feature_proxy::random_value;
 use crate::segment_evaluation::TargetingRules;
 
 use crate::errors::Result;
+use std::io::Cursor;
 
 /// Provides a snapshot of a [`Feature`].
 #[derive(Debug)]
@@ -129,6 +131,15 @@ impl Feature for FeatureSnapshot {
         let value = self.get_value(entity)?;
         value.try_into()
     }
+}
+
+pub(crate) fn random_value(v: &str) -> u32 {
+    let max_hash = u32::MAX;
+    (f64::from(hash(v)) / f64::from(max_hash) * 100.0) as u32
+}
+
+fn hash(v: &str) -> u32 {
+    murmur3_32(&mut Cursor::new(v), 0).expect("Cannot hash the value.")
 }
 
 #[cfg(test)]
