+++ src/tests/test_get_feature.rs
@@ -16,7 +16,7 @@ use std::collections::HashMap;
 
 use crate::network::models::ConfigurationJson;
 
-use crate::client::configuration::Configuration;
+use crate::models::Configuration;
 use crate::{AppConfigurationClient, ConfigurationProvider};
 use crate::{AppConfigurationOffline, Value};
 use rstest::*;
